Merges 61 upstream commits: array unification (6f6b79fd), always-use
prepared calls (150f2864), immutable obj core list (e87511f4, brings
im-rc Vectors), unified ContextBuilder (a486673b, deletes map.rs),
rowan unary wrapping (66126184), function folder Builtins abstraction
(5e8cbcdb), and — pulled in from the originally-planned stage 5 —
upstream's static analysis (1979dbda + d5225b82), which is the designed
replacement for the fork's UsedVars trimming retired at stage 2.

Note: the originally planned checkpoint 259b3abb does not compile
standalone (upstream mid-refactor breakage between the analyze split
commits); the stage was retargeted to d5225b82, the first buildable
commit after, verified in a clean worktree.

IMPORTANT pin change: jrsonnet-gcmodule switched from the grafana
leak-fix fork (v0.4.1 branch, no im-rc support) to crates.io 0.4.4
with the im-rc feature, required by upstream's new obj internals.
The leak fix MUST be rebased and restored before this branch is PRed
(tracked; grafana/jrsonnet-gcmodule branch push needed).

Fork features re-applied onto upstream's rewritten core:
- ArrValue::extended always-flatten (adapted to unified array API:
  Vec<Thunk<Val>> via From, u32 len, Option overflow contract)
- Go-style number concat in operator.rs; %s go-float hook in format.rs;
  go-floats machinery in manifest.rs (upstream BlackBoxFormat kept)
- stdlib: Builtins folder adopted, but manifestYamlDoc/Stream removed
  from it in favor of the fork's settings-carrying registrations;
  regex natives (incl. new regexQuoteMeta) always enabled, exp-regex
  feature gate dropped; CachedUnbound weak-key pruning in val.rs
- jrsonnet-lint: unary operand walker updated for upstream's new
  EXPR-wrapped unary shape (66126184), handling both shapes
- rtk: builtin registration migrated from ::INST to struct-init
  pattern, u32 array len cast
- tests: adopted upstream's direct serde_json::Value comparison
  (json_structural_diff dep dropped upstream), kept fork's graceful
  missing-testsuite skip
- Re-blessed one analysis snapshot upstream itself left stale at this
  commit (analyze_tests -> analysis_tests rename); marked upstream's
  Closures doc example as text (doctest didn't compile)

Validated: build clean, make test green, make check-golden-fixtures
byte-identical, issue #30 repro ms-fast at depth 100,
super_assertion_recursion repro OK.

+++ cmds/rtk/src/commands/validate/common.rs
@@ -248,7 +248,7 @@ pub fn eval_jsonnet_snippet_array_pooled(
 			}
 		};
 
-		let mut out: Vec<ElementResult> = Vec::with_capacity(arr.len());
+		let mut out: Vec<ElementResult> = Vec::with_capacity(arr.len() as usize);
 		for idx in 0..arr.len() {
 			let elem_result: ElementResult = match arr.get(idx) {
 				Ok(Some(val)) => match val.manifest(JsonFormat::default()) {
